spliting GenericSettings class into its own file
